% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupled2_rwmh_kernel.R
\name{coupled2_rwmh_kernel}
\alias{coupled2_rwmh_kernel}
\title{RWM kernel for two coupled chains}
\usage{
coupled2_rwmh_kernel(
  level,
  state1,
  state2,
  identical,
  tuning,
  proposal_coupling
)
}
\arguments{
\item{level}{an integer that determines probability distribution}

\item{state1}{a list with state of the first chain}

\item{state2}{a list with state of the second chain}

\item{identical}{a boolean variable that is True if chains are identical and False otherwise}

\item{tuning}{a list of parameters neeeded for RWM (standatd deviation)}

\item{proposal_coupling}{function that determing proposal coupling}
}
\value{
a list that contains states of two chains, updated value of the flag "identical" and cost of computations
}
\description{
generation of coupled proposals + accept/reject step
}
