% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupled2_rwmh_kernel_ext.R
\name{coupled2_rwmh_kernel_ext}
\alias{coupled2_rwmh_kernel_ext}
\title{RWM kernel for two chains that target distributions at different levels with synchronous couling of pCN proposals}
\usage{
coupled2_rwmh_kernel_ext(level, state1, state2, tuning1, tuning2)
}
\arguments{
\item{level}{an integer that determines probability distribution}

\item{state1}{a list with state of the first chain}

\item{state2}{a list with state of the second chain}

\item{tuning}{a list of parameters neeeded for RWM (standatd deviation)}
}
\value{
a list that contains states of two chains and cost of computations
}
\description{
generation of coupled proposals + accept/reject step
}
