% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupled4_pcn_kernel.R
\name{coupled4_pcn_kernel}
\alias{coupled4_pcn_kernel}
\title{pCN kernel for four coupled chains}
\usage{
coupled4_pcn_kernel(
  level,
  state_coarse1,
  state_coarse2,
  state_fine1,
  state_fine2,
  identical_coarse,
  identical_fine,
  tuning_coarse,
  tuning_fine,
  proposal_coupling
)
}
\arguments{
\item{level}{an integer that determines probability distribution: level == l}

\item{state_coarse1}{a list with state of the first chain at the level "l - 1"}

\item{state_coarse2}{a list with state of the second chain at the level "l - 1"}

\item{state_fine1}{a list with state of the first chain at the level "l"}

\item{state_fine2}{a list with state of the second chain at the level "l"}

\item{identical_coarse}{a boolean variable that is True if chains at the level "l - 1" are identical and False otherwise}

\item{identical_fine}{a boolean variable that is True if chains at the level "l" are identical and False otherwise}

\item{tuning_coarse}{a list of parameters neeeded for proposal generation with pCN (aka standard deviation and rho) for chains at the level "l - 1"}

\item{tuning_fine}{a list of parameters neeeded for proposal generation with pCN (aka standard deviation and rho) for chains at the level "l"}

\item{proposal_coupling}{function that determing proposal coupling}
}
\value{
a list that contains states all four chains, updated value of the flag "identical_coarse" and "identical_coarse", cost of computations
}
\description{
generation of coupled proposals + accept/reject step
}
