% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxprocess_model.R
\name{coxprocess_model}
\alias{coxprocess_model}
\title{Construct log-Gaussian Cox process model}
\usage{
coxprocess_model(parameters, dataset, level)
}
\arguments{
\item{parameters}{a list of model parameters (sigmasq, mu, beta)}

\item{dataset}{a list of observed spatial locations}

\item{level}{an integer that determines the spatial resolution}
}
\value{
a list containing the keys \code{dimension}, \code{ngrid}, \code{logtarget}, \code{gradlogtarget}, \code{rinit} and \code{gradient_parameters}
}
\description{
Construct log-Gaussian Cox process model at a given spatial resolution
}
