% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxprocess_kernels.R
\name{coxprocess_multilevel_kernels}
\alias{coxprocess_multilevel_kernels}
\title{Multilevel Markov chain Monte Carlo kernels for log-Gaussian Cox process model}
\usage{
coxprocess_multilevel_kernels(model_coarse, model_fine, tuning)
}
\arguments{
\item{model_coarse}{list for coarser resolution model containing \code{logtarget}, \code{gradlogtarget} and \code{dimension}}

\item{tuning}{list containing \code{stepsize} and \code{nsteps}}

\item{model_finer}{list for finer resolution model containing \code{logtarget}, \code{gradlogtarget} and \code{dimension}}
}
\value{
a list containing the keys \code{single_kernel}, \code{coupled2_kernel}
}
\description{
This function defines multilevel Markov chain Monte Carlo kernels
}
