% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxprocess_estimators.R
\name{coxprocess_unbiased_expectation}
\alias{coxprocess_unbiased_expectation}
\title{Unbiased estimator for the value at level zero}
\usage{
coxprocess_unbiased_expectation(
  parameters,
  model,
  mcmc,
  h = function(x) x,
  k = 0,
  m = 1,
  max_iterations = Inf
)
}
\arguments{
\item{parameters}{a list of model parameters (sigmasq, mu, beta)}

\item{model}{a list of model functions with keys \code{logtarget}, \code{gradlogtarget}, \code{rinit} and \code{dimension}}

\item{mcmc}{a list of MCMC kernels with keys \code{single_kernel} and \code{coupled2_kernel}}

\item{h}{function that represents quantity of interest}

\item{k}{an integer: lower bound for time-averaging}

\item{m}{an integer:  upper bound for time-averaging}

\item{max_iterations}{iteration at which to stop the while loop (default to infinity)}
}
\value{
a list with the value of MCMC estimator without correction, value of Unbiased MCMC estimator, meeting time, value of iteration counter, flag that is "True" if chains have met before the iteration counter reached the value in max_iterations, cost of calculations
}
\description{
Generate two coupled chains and compute the estimator
}
