% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxprocess_estimators.R
\name{coxprocess_unbiased_increment}
\alias{coxprocess_unbiased_increment}
\title{Unbiased estimator for the value of the increment at the given level}
\usage{
coxprocess_unbiased_increment(
  parameters,
  model_coarse,
  model_fine,
  mcmc,
  h = function(x) x,
  k = 0,
  m = 1,
  max_iterations = Inf
)
}
\arguments{
\item{parameters}{a list of model parameters (sigmasq, mu, beta)}

\item{model_coarse}{a list of functions for coarser resolution model with keys \code{logtarget}, \code{gradlogtarget}, \code{rinit} and \code{dimension}}

\item{model_fine}{a list of functions for finer resolution model with keys \code{logtarget}, \code{gradlogtarget}, \code{rinit} and \code{dimension}}

\item{mcmc}{a list of MCMC kernels with keys \code{multilevel_kernel} and \code{coupled4_kernel}}

\item{h}{function that represents quantity of interest. Depends on level and spatial argument}

\item{k}{an integer: lower bound for time-averaging}

\item{m}{an integer:  upper bound for time-averaging}

\item{max_iterations}{iteration at which to stop the while loop (default to infinity)}
}
\value{
a list with the value of MCMC estimator without correction, value of Unbiased MCMC estimator, meeting time, value of iteration counter, flag that is "True" if chains have met before the iteration counter reached the value in max_iterations, cost of calculations
}
\description{
Generate four coupled chains and compute the estimator of the increment
}
