% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leapfrog.R
\name{leapfrog}
\alias{leapfrog}
\title{Leapfrog inregration for HMC kernel}
\usage{
leapfrog(level, x, v, tuning)
}
\arguments{
\item{level}{an integer that controls probability distribution utilized for calculation of gradients of the potential}

\item{x}{position of the particle}

\item{v}{velocity of the particle}

\item{tuning}{a list that contains the stepsize and number of steps for the leapfrog integration}
}
\value{
list that contains new position of the particle, new velocsity of the particle and the cost of leapfrog integration (proportional to number of gradient evaluation)
}
\description{
leapfrog integrtion
}
