% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximal2_rwmh_coupling.R
\name{maximal2_rwmh_coupling}
\alias{maximal2_rwmh_coupling}
\title{maximal coupling of 2 RWM proposals}
\usage{
maximal2_rwmh_coupling(chain_state1, chain_state2, identical, tuning)
}
\arguments{
\item{chain_state1}{a vector with coordinates of the first particle}

\item{chain_state2}{a vector with coordinates of the second particle}

\item{identical}{a flag that is True if chains are identical and False otherwise}

\item{tuning}{a list that contains parameters needed for RWM: standard devation}
}
\value{
a list that contains state of the first chain, state of the second chain, updated value of the flag "identical", cost of proposal generation
}
\description{
Generate pCn proposals for two chains via maximal coupling
}
