% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximal4_pcn_independent_coupling.R
\name{maximal4_pcn_independent_coupling}
\alias{maximal4_pcn_independent_coupling}
\title{independent maximal coupling of 4 pCN proposals}
\usage{
maximal4_pcn_independent_coupling(
  chain_state_coarse1,
  chain_state_coarse2,
  chain_state_fine1,
  chain_state_fine2,
  identical_coarse,
  identical_fine,
  tuning_coarse,
  tuning_fine
)
}
\arguments{
\item{chain_state_coarse1}{a vector with coordinates of the first particle at the coarse level}

\item{chain_state_coarse2}{a vector with coordinates of the second particle at the coarse level}

\item{chain_state_fine1}{a vector with coordinates of the first particle at the fine level}

\item{chain_state_fine2}{a vector with coordinates of the second particle at the fine level}

\item{identical_coarse}{a flag that is True if chains at the coarse level are identical (coincide) and False otherwise}

\item{identical_fine}{a flag that is True if chains at the fine level are identical (coincide) and False otherwise}

\item{tuning_coarse}{a list that contains parameters needed for genertaion of pCN proposal at the fine level: standard devation and rho}
}
\value{
a list that contains state of the first chain at the coarse level, state of the second chain at the coarse level, state of the second chain, updated values of flags "identical_coarse" and "identical_fine", cost of proposal generation
}
\description{
Generate pCN proposals for two chains via maximal coupling for pairs of chains at the coarse and fine levels independently
}
