% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture2_hmc_kernel_extra_level.R
\name{mixture2_hmc_kernel_extra_level}
\alias{mixture2_hmc_kernel_extra_level}
\title{Mixture of kernels with HMC synchronous coupling and RWM maximal coupling for two coupled chains at two different levels}
\usage{
mixture2_hmc_kernel_extra_level(
  level,
  state1,
  state2,
  identical,
  tuning,
  proposal_coupling
)
}
\arguments{
\item{level}{an integer that controls the probability distribution in a multi-level setting}

\item{state1}{a list that contains vector of coordinates and log of the density of probability distribution for the chain at the level "level - 1"}

\item{state2}{a list that contains vector of coordinates and log of the density of probability distribution for the chain at the level "level"}

\item{identical}{a boolean variable: "True" if chains coincide, "False" otherwise. In the setting of two chains that target different levels, this parameter is always "False". It is kept only for compatibility with other kernels and couplings}

\item{tuning}{list of parameters for HMC and RWM kernels (number os steps and stepsize for HMC, standard devation for RWM)}
}
\value{
list that contains updated states of the first and the second chain, updated value of the flag "identical" and cost of computations
}
\description{
Either HMC kernel or RWM kernel is selected for MCMC iteration with a certain probability
}
