% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture4_hmc_kernel.R
\name{mixture4_hmc_kernel}
\alias{mixture4_hmc_kernel}
\title{Mixture of kernels with HMC synchronous coupling and RWM maximal coupling for four coupled chains}
\usage{
mixture4_hmc_kernel(
  level,
  state_coarse1,
  state_coarse2,
  state_fine1,
  state_fine2,
  identical_coarse,
  identical_fine,
  tuning_coarse,
  tuning_fine,
  proposal_coupling
)
}
\arguments{
\item{level}{an integer that controls the probability distribution in a multi-level setting}

\item{state_coarse1}{a list that contains vector of coordinates and log of the density of probability distribution for the first chain at the coarse level: "level - 1"}

\item{state_coarse2}{a list that contains vector of coordinates and log of the density of probability distribution for the second chain at the coarse level: "level - 1"}

\item{state_fine1}{a list that contains vector of coordinates and log of the density of probability distribution for the first chain at the fine level: "level"}

\item{state_fine2}{a list that contains vector of coordinates and log of the density of probability distribution for the second chain at the fine level: "level"}

\item{identical_coarse}{a boolean variable: "True" if chains at the coarse level coincide, "False" otherwise}

\item{identical_fine}{a boolean variable: "True" if chains at the fine level coincide, "False" otherwise}

\item{tuning_coarse}{list of parameters for HMC and RWM kernels (number os steps and stepsize for HMC, standard devation for RWM) for two chains at the coarse level}

\item{tuning_fine}{list of parameters for HMC and RWM kernels (number os steps and stepsize for HMC, standard devation for RWM) for two chains at the fine level}
}
\value{
list that contains updated states of the first and the second chains at the coarse level,  updated states of the first and the second chains at the fine level, updated value of the flag "identical_coarse", updated value of the flag "identical_fine" and cost of computations
}
\description{
Either HMC kernel or RWM kernel is selected for MCMC iteration with a certain probability
}
