% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture4_pcn_kernel.R
\name{mixture4_pcn_kernel}
\alias{mixture4_pcn_kernel}
\title{Mixture of kernels with pCn synchronous coupling and pCN maximal coupling for four coupled chains}
\usage{
mixture4_pcn_kernel(
  level,
  state_coarse1,
  state_coarse2,
  state_fine1,
  state_fine2,
  identical_coarse,
  identical_fine,
  tuning_coarse,
  tuning_fine,
  proposal_coupling
)
}
\arguments{
\item{level}{an integer that controls the probability distribution in a multi-level setting}

\item{state_coarse1}{a list that contains vector of coordinates and log of the density of probability distribution for the first chain at the coarse level: "level - 1"}

\item{state_coarse2}{a list that contains vector of coordinates and log of the density of probability distribution for the second chain at the coarse level: "level - 1"}

\item{state_fine1}{a list that contains vector of coordinates and log of the density of probability distribution for the first chain at the fine level: "level"}

\item{state_fine2}{a list that contains vector of coordinates and log of the density of probability distribution for the second chain at the fine level: "level"}

\item{identical_coarse}{a boolean variable: "True" if chains at the coarse level coincide, "False" otherwise}

\item{identical_fine}{a boolean variable: "True" if chains at the fine level coincide, "False" otherwise}

\item{tuning_coarse}{a list of parameters for pCN kernel (standard devation and rho) for two chains at the coarse level}

\item{tuning_fine}{a list of parameters for pCN kernel (standard devation and rho) for two chains at the fine level}
}
\value{
list that contains updated states of the first and the second chains at the coarse level,  updated states of the first and the second chains at the fine level, updated value of the flag "identical_coarse", updated value of the flag "identical_fine" and cost of computations
}
\description{
Either kernel with pCN synchronous coupling or kernel with pCN maxiimal coupling is selected for MCMC iteration with a certain probability
}
