% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_hmc_kernel.R
\name{single_hmc_kernel}
\alias{single_hmc_kernel}
\title{HMC kernel for a single MCMC chain that target probability distribution at the given level}
\usage{
single_hmc_kernel(level, state, tuning)
}
\arguments{
\item{level}{a integer that determines density of probability distribution inthe multi-level approach}

\item{state}{a list with current position of the particle (element of the chain) and log of the density of probability distribution}

\item{tuning}{a list of parameters for HMC iteration: stepsize and number of steps}
}
\value{
a updated state of the chain and cost of computations
}
\description{
Generation of HMC proposal vil leapfrog integration and accept/reject step
}
