% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_mixture_hmc_kernel.R
\name{single_mixture_hmc_kernel}
\alias{single_mixture_hmc_kernel}
\title{Mixture of kernels with HMC synchronous coupling and RWM maximal coupling a single chain at the given level}
\usage{
single_mixture_hmc_kernel(level, state, tuning)
}
\arguments{
\item{level}{an integer that controls the probability distribution in a multi-level setting}

\item{state}{a list that contains vector of coordinates and log of the density of probability distribution for the first chain}

\item{tuning}{list of parameters for HMC and RWM kernels (number os steps and stepsize for HMC, standard devation for RWM)}
}
\value{
list that contains updated states of the first and the second chain and cost of computations
}
\description{
Either HMC kernel or RWM kernel is selected for MCMC iteration with a certain probability
}
