% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchronous2_hmc_coupling.R
\name{synchronous2_hmc_coupling}
\alias{synchronous2_hmc_coupling}
\title{Syncronous coupling of two HMC proposals}
\usage{
synchronous2_hmc_coupling(level, chain_state1, chain_state2, identical, tuning)
}
\arguments{
\item{level}{an integer number that determines probability distribution in the multi-level approach}

\item{chain_state1}{a vector with the position of the first particle}

\item{chain_state2}{a vector with the position of the second particle}

\item{identical}{a boolean variable takes the value True if two chains coincide and False otherwise}

\item{tuning}{a list that contains parameters required for leapfrog integration: stepsize and number os steps}
}
\value{
a list with updated values of the states of the first and the second chain, velocites of each of the particles, initial velocity (random sample from normal distribution), updated value of the "identical" flag and cost of computations
}
\description{
Sampling of common momentum for both particles and leapfrog integration
}
