% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchronous2_rwmh_coupling.R
\name{synchronous2_rwmh_coupling}
\alias{synchronous2_rwmh_coupling}
\title{Synchronous coupling of RWM proposals for two chains}
\usage{
synchronous2_rwmh_coupling(chain_state1, chain_state2, identical, tuning)
}
\arguments{
\item{chain_state1}{a vector with position of the first particle}

\item{chain_state2}{a vector with position of the second particle}

\item{identical}{a boolean variable, that takes value "True" if particles are at the same position and}

\item{tuning}{a list that contains parameters required for RWM: standard deviation}
}
\value{
list that contains updated states of the first and the second particles, updatd value of identical flag and the cost of calculations
}
\description{
Generation of two RWM proposals based on the common Gaussian increment
}
