% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbiased_expectation.R
\name{unbiased_expectation}
\alias{unbiased_expectation}
\title{Unbiased estimator for the value at level zero}
\usage{
unbiased_expectation(
  level,
  rinit,
  single_kernel,
  tuning,
  coupled_kernel,
  proposal_coupling,
  h = function(l, x) x,
  k = 0,
  m = 1,
  max_iterations = Inf
)
}
\arguments{
\item{level}{an integer that determines the target probability distribution of two chains}

\item{rinit}{function that is utilized for initialization of chains (for instance, samples from prior)}

\item{single_kernel}{function that makes a single step through a specified MCMC kernel}

\item{tuning}{a list of parameters reuired for MCMC iterations (for instance standard deviation for RWM)}

\item{coupled_kernel}{function that makes a single step through a specified coupled MCMC kernel (for 2 chains)}

\item{proposal_coupling}{function that generates proposal for a given input}

\item{h}{function that represents quantity of interest. Depends on level and spatial argument.}

\item{k}{an integer: lower bound for time-averaging}

\item{m}{an integer:  upper bound for time-averaging}

\item{max_iterations}{and integer: bound for the number of steps through coupled MCMC kernel}
}
\value{
a list with the value of MCMC estimator without correction, value of Unbiased MCMC estimator, meeting time, value of iteration counter, flag that is "True" if chains have met before the iteration counter reached the value in max_iterations, cost of calculations
}
\description{
Generate two coupled chains and compute the estimator
}
