% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbiased_increment.R
\name{unbiased_increment}
\alias{unbiased_increment}
\title{Unbiased estimator for the value of the increment at the given level}
\usage{
unbiased_increment(
  level,
  rinit,
  single_kernel,
  coupled2_kernel,
  coupled4_kernel,
  proposal_coupling2,
  proposal_coupling4,
  tuning,
  tuning_coarse,
  tuning_fine,
  h = function(l, x) x,
  k = 0,
  m = 1,
  max_iterations = Inf,
  sampling_factor = 0.2
)
}
\arguments{
\item{level}{an integer that determines the target probability distributions of four chains}

\item{rinit}{function that is utilized for initialization of chains (for instance, samples from prior)}

\item{single_kernel}{a function that makes a single step through a specified MCMC kernel for a single chain}

\item{coupled2_kernel}{a function that makes a single step through a specified coupled MCMC kernel for all four chains}

\item{proposal_coupling2}{a function that generates proposal for a given input for two coupled chains (burn-in + lag)}

\item{proposal_coupling4}{a function that generates proposal for a given input for four coupled chains (main run)}

\item{tuning}{a list of parameters required for MCMC iterations (for instance standard deviation for RWM)}

\item{tuning_coarse}{a list of parameters required for MCMC iterations for the coarse level (for instance standard deviation for RWM)}

\item{tuning_fine}{a list of parameters required for MCMC iterations for the fine level (for instance standard deviation for RWM)}

\item{h}{function that represents quantity of interest. Depends on level and spatial argument.}

\item{k}{an integer: lower bound for time-averaging}

\item{m}{an integer:  upper bound for time-averaging}

\item{max_iterations}{iteration at which to stop the while loop (default to infinity)}

\item{samping_factor}{a real value that controls the magnitude of perturbation at the initialization step}
}
\value{
a list with the value of MCMC estimator without correction, value of Unbiased MCMC estimator, meeting time, value of iteration counter, flag that is "True" if chains have met before the iteration counter reached the value in max_iterations, cost of calculations
}
\description{
Generate four coupled chains and compute the estimator of the increment
}
