% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnbiasedMultilevel.R
\docType{package}
\name{UnbiasedMultilevel}
\alias{UnbiasedMultilevel}
\title{UnbiasedMultilevel package}
\description{
Unbiased estimators
}
\details{
This R package provides unbiased estimators
}
\author{
Jeremy Heng <jeremyhengjm@gmail.com>
}
\keyword{package}
